ocr/requirements.txt
tensorflow==2.10.0
numpy==1.26.4
pillow==9.4.0
flask==3.1.0
pytesseract==0.3.13
gunicorn==20.1.0